+++ CopilotKit/packages/runtime/CHANGELOG.md
@@ -1,5 +1,65 @@
 # @copilotkit/runtime
 
+## 1.10.4
+
+### Patch Changes
+
+- a640d8e: - feat: update latest agui langgraph for subgraphs support
+  - feat: update latest agui core packages
+- Updated dependencies [a640d8e]
+  - @copilotkit/shared@1.10.4
+
+## 1.10.4-next.3
+
+### Patch Changes
+
+- @copilotkit/shared@1.10.4-next.3
+
+## 1.10.4-next.2
+
+### Patch Changes
+
+- @copilotkit/shared@1.10.4-next.2
+
+## 1.10.4-next.1
+
+### Patch Changes
+
+- a640d8e: - feat: update latest agui langgraph for subgraphs support
+  - feat: update latest agui core packages
+- Updated dependencies [a640d8e]
+  - @copilotkit/shared@1.10.4-next.1
+
+## 1.10.4-next.0
+
+### Patch Changes
+
+- @copilotkit/shared@1.10.4-next.0
+
+## 1.10.3
+
+### Patch Changes
+
+- ea74047: - fix: surface run errors from agui
+- a7bb2f0: - fix: fix how node names are read to support newer langgraph versions
+- 21e12af: - feat: send streamSubgraphs forwarded prop to AGUI
+- Updated dependencies [ea74047]
+  - @copilotkit/shared@1.10.3
+
+## 1.10.3-next.3
+
+### Patch Changes
+
+- 21e12af: - feat: send streamSubgraphs forwarded prop to AGUI
+  - @copilotkit/shared@1.10.3-next.3
+
+## 1.10.3-next.2
+
+### Patch Changes
+
+- a7bb2f0: - fix: fix how node names are read to support newer langgraph versions
+  - @copilotkit/shared@1.10.3-next.2
+
 ## 1.10.3-next.1
 
 ### Patch Changes